pyproject.toml
[tool.poetry]
name = "pokehelper"
version = "0.1.0"
description = ""
authors = ["Kyle He <keelayhe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
motor = "^2.3.0"
jishaku = "^1.19.2"
discord-ext-menus = {git = "https://github.com/oliver-ni/discord-ext-menus.git"}
dnspython = "^2.0.0"
discord-ext-events = {git = "https://github.com/Ext-Creators/discord-ext-events.git"}
durations-nlp = "^1.0.1"
umongo = "^3.0.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
